Get container exit code reliably
This is a kludge to work around a Docker [exit code bug](https://github.com/kubernetes/kubernetes/issues/41516).

Now, when docker_termination_logging is set to "true" for the scheduled workflow at submission time, a message in the termination log will be expected for determination of the container exit status.