fix execution id creation & submission race
Create the execution id in the prepare state and add it to the run state
as part of the submit event. This eliminates the race between k8s
submission and execution id persistence that caused the runner to not be
able to look up run states for newly started pods.

For backwards compatibility, the `executionId` field is kept in the
`submitted` event to ensure that we can replay old events and do not
lose execution id's when deploying this change.