cli: fix slow startup
Turns out logback tries to look up the host name of the machine by
making a DNS query. That is usually quick on machines in a datacenter
but often fails (slowly) on workstations.

By replacing logback with slf4j-simple in the cli we can eliminate
the above host name lookup.