Fix bug in positioning the cleanup within the state machine.
Also fix failing tests.

There are instances where the a workflowInstance's runState is still kept
in the stateManager even though the execution failed. For example, the runState is
removed from the stateManager only when the runState isTerminal (meaning, Done or Error),
whereas Failed and Terminated are not considered terminal states. In
the case of cleaning up pods, we want to clean the pods once a workflowInstance
enters one of all those 4 states (Failed/Terminated/Done/Error). Running cleanup(..) only
under the condition !runState.isPresent() prevented us to do so.

Now we try to cleanup the pod no matter of whether the runState isPresent or not in the
stateManager. The cleanup method checks whether **the pod** is terminated, and then
decides whether to clean it up.