racing between poller and watcher
We have observed cases that pod is not seen for running state. That
happens after poller has fetched a list of pods, then styx creates a
new pod, that pod transits into running and the state machine transits
into running as well. So we end up with having a running state but no
pod.

To fix this, first get a snapshot of running states and then fetch pod
list. If the state is running, the associated pod should be in the list;
while if the state is not running yet, we will not take any action.