Group shards in async transactions when initializing
The shards are all unique so there won't be any conflicts between different transactions.
Why groups of 15? It seems to yield fastest results. If it's larger (say the max of 25 entities per transactions),
the overhead lies in looping through the shards and adding them to the transactions;
if it's smaller, the overhead lies in creating the transactions themselves.