strongly consistent wfi listing using shards
Active workflow instances are indexed using 128 entity groups (index
shards). 

The id's of the instances are stored as the names of child entities in
the index shards. The entities themselves are empty.

This avoids the 1MB max entity size limit that might be hit if instance
id's are stored as list properties.