Please error prone on date handling in validation
We validate that offset by trying to apply it to a date.
Doing that we just created a now date but error prone complained
that we were relying on the default timezone instead of specifying one.
It is true but it doesn't matter. In any case the fix is simple